NSFWImageClassification/requirements.txt
accelerate==0.26.0
huggingface-hub==0.30.1
numpy==2.1.1
opencv-python==4.11.0.86
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
transformers==4.50.3
ultralytics==8.3.100
pillow==11.1.0
